개인 프로젝트 final

+++ src/main/webapp/resources/js/reply.js
@@ -56,20 +56,22 @@ $(function(){
 					    </div>
 					    <div class="comment_wrap">
 					      <div class="comment_info">
-					        <span class="userName badge badge-pill badge-info mr-2">홍길동</span>
+					        <span class="userName badge badge-pill badge-info mr-2">${elem.replyer}</span>
 					        <span class="badge badge-dark">${elem.replyDate}</span>
 					      </div>
 					      <div class="comment_content py-2">${elem.reply}</div>
 					    </div>
-					  </div>
-					  <div class="reply_modify">
-					    <button type="button" class="btn btn-light dropdown-toggle" data-toggle="dropdown">변경</button>
-					    <div class="dropdown-menu">						   
-					      <a class="dropdown-item" href="modify">수정</a>
-					      <a class="dropdown-item" href="delete">삭제</a>
-					    </div>
-					  </div>
-					 </div>
+					  </div>`
+					  if(memberId==elem.replyer || auth.includes('ROLE_ADMIN')){
+						  replyList += `<div class="reply_modify">
+						    <button type="button" class="btn btn-light dropdown-toggle" data-toggle="dropdown">변경</button>
+						    <div class="dropdown-menu">						   
+						      <a class="dropdown-item" href="modify">수정</a>
+						      <a class="dropdown-item" href="delete">삭제</a>
+						    </div>
+						  </div>`					  
+					  }
+					replyList+= `</div>
 					</li>`				
 			});
 			replyContainer.html(replyList);
@@ -110,6 +112,12 @@ $(function(){
 		e.preventDefault();// a태그  기본동작 금지
 		let rno = $(this).closest('li').data('rno'); // 댓글 번호 가져오기
 		let operation = $(this).attr('href');// 수정/삭제 동작 결정
+		let replyer = $(this).closest('li').find('.userName').text();
+
+		// 작성자가 일치하지 않거나 관리자가 아니면
+		if(replyer!=memberId && !auth.includes('ROLE_ADMIN')){
+			return;
+		}
 		
 		if(operation=='delete'){ // 삭제 처리 
 			replyService.remove(rno,function(result){
